deleting a user's advertisements and clearing them from other users' trash cans
When a user is deleted, an event is sent to the AdvertisementService to delete all of their ads.

After the ads are deleted, an event is sent back to the UserService.

In UserService, ads are removed from the garbage cans of all users who had them.

Added case handling if the user has already been deleted, but a dumpster cleanup and notification is required.